+++ lib/views/home/home_screen.dart
@@ -1,11 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../models/tab_item.dart';
 import '../app_localizations.dart';
-import '../widgets/tabs_screen.dart';
-import 'notices_page.dart';
-import 'groups_page.dart';
+import '../widgets/section_header.dart';
+import 'notices_section.dart';
+import 'groups_section.dart';
 
 class HomeScreen extends HookConsumerWidget {
   const HomeScreen({super.key});
@@ -14,18 +13,18 @@ class HomeScreen extends HookConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final t = AppLocalizations.of(context)!;
 
-    return TabsScreen(
-      items: [
-        TabItem(
-          icon: Icons.notifications,
-          label: t.notices,
-          page: const NoticesPage(),
+    return CustomScrollView(
+      slivers: [
+        SectionHeader(
+          title: t.notices,
+          leading: Icons.notifications,
         ),
-        TabItem(
-          icon: Icons.people,
-          label: t.groups,
-          page: const GroupsPage(),
+        const NoticesSection(),
+        SectionHeader(
+          title: t.groups,
+          leading: Icons.people,
         ),
+        const GroupsSection(),
       ],
     );
   }
