--- lib/views/home/notices_page.dart
@@ -1,18 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-class NoticesPage extends HookConsumerWidget {
-  const NoticesPage({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return const SingleChildScrollView(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(''),
-        ],
-      ),
-    );
-  }
-}
