--- lib/views/me/preferences_page.dart
@@ -1,24 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-import '../../config.dart';
-import '../widgets/select_site.dart';
-
-class PreferencesPage extends HookConsumerWidget {
-  const PreferencesPage({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return const SingleChildScrollView(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Padding(
-            padding: cardItemPadding,
-            child: SelectSite(),
-          ),
-        ],
-      ),
-    );
-  }
-}
